requirements.txt
neo4j==5.28.1
python-dotenv==1.1.0
pandas==2.2.3
streamlit==1.40.1
plotly==5.24.1
